Replace permission asserts with HTTP 403 Forbidden

+++ coursereg/views/participants.py
@@ -9,25 +9,27 @@ from coursereg import models
 from django.core.mail import send_mail
 from django.db.models import Q
 from django.conf import settings
+from django.core.exceptions import PermissionDenied
+from django.views.decorators.http import require_POST
 
+@require_POST
 @login_required
 def create(request):
-    assert request.method == 'POST'
     course_id = request.POST['course_id']
     user_id = request.POST['user_id']
     reg_type = request.POST['reg_type']
 
     user = models.User.objects.get(id=user_id)
     if request.POST['origin'] == 'adviser':
-        assert request.user == user.adviser
+        if not request.user == user.adviser: raise PermissionDenied
     else:
-        assert request.user == user
+        if not request.user == user: raise PermissionDenied
 
     if not course_id:
         messages.error(request, 'Select a course.')
     else:
         course = models.Course.objects.get(id=course_id)
-        assert not course.is_last_reg_date_passed()
+        if course.is_last_reg_date_passed(): raise PermissionDenied
         if models.Participant.objects.filter(user__id=user_id, course__id=course_id):
             messages.error(request, 'Already registered for %s.' % course)
         else:
@@ -57,22 +59,23 @@ def create(request):
 
     return redirect(request.POST.get('next', reverse('coursereg:index')))
 
+@require_POST
 @login_required
 def update(request, participant_id):
-    assert request.method == 'POST'
     participant = models.Participant.objects.get(id=participant_id)
     if request.POST['origin'] == 'instructor':
-        assert models.Participant.objects.filter(course=participant.course,
-            user=request.user, participant_type=models.Participant.PARTICIPANT_INSTRUCTOR)
+        if not models.Participant.objects.filter(course=participant.course,
+                user=request.user, participant_type=models.Participant.PARTICIPANT_INSTRUCTOR):
+            raise PermissionDenied
         if request.POST['action'] == 'approve':
-            assert participant.is_adviser_approved
+            if not participant.is_adviser_approved: raise PermissionDenied
             participant.is_instructor_approved = True
             participant.save()
             student = participant.user
             student.is_dcc_review_pending = True
             student.save()
         elif request.POST['action'] == 'reject':
-            assert not participant.is_instructor_approved
+            if participant.is_instructor_approved: raise PermissionDenied
             msg = 'Application for %s has been rejected by the course instructor.' % participant.course
             models.Notification.objects.create(user=participant.user,
                                                origin=models.Notification.ORIGIN_INSTRUCTOR,
@@ -83,22 +86,22 @@ def update(request, participant_id):
             except:
                 messages.warning(request, 'Error sending e-mail. But a notification has been created on this website.')
         elif request.POST['action'] == 'grade':
-            assert not participant.course.is_last_grade_date_passed()
+            if participant.course.is_last_grade_date_passed(): raise PermissionDenied
             participant.grade = models.Grade.objects.get(id=request.POST['grade'])
             participant.save()
     elif request.POST['origin'] == 'adviser':
-        assert participant.user.adviser == request.user
+        if not participant.user.adviser == request.user: raise PermissionDenied
         if request.POST['action'] == 'state_change':
-            assert not participant.course.is_drop_date_passed()
+            if participant.course.is_drop_date_passed(): raise PermissionDenied
             state = request.POST['state']
             if state == 'credit':
                 if not participant.is_credit:
-                    assert not participant.course.is_last_conversion_date_passed()
+                    if participant.course.is_last_conversion_date_passed(): raise PermissionDenied
                 participant.is_credit = True
                 participant.is_drop = False
             elif state == 'audit':
                 if participant.is_credit:
-                    assert not participant.course.is_last_conversion_date_passed()
+                    if participant.course.is_last_conversion_date_passed(): raise PermissionDenied
                 participant.is_credit = False
                 participant.is_drop = False
             elif state == 'drop':
@@ -123,7 +126,7 @@ def update(request, participant_id):
             student.is_dcc_review_pending = True
             student.save()
         elif request.POST['action'] == 'delete':
-            assert not participant.course.is_last_reg_date_passed() or not participant.is_instructor_approved
+            if participant.course.is_last_reg_date_passed() and participant.is_instructor_approved: raise PermissionDenied
             msg = 'Application for %s has been rejected by your adviser.' % participant.course
             models.Notification.objects.create(user=participant.user,
                                                origin=models.Notification.ORIGIN_ADVISER,
@@ -135,12 +138,12 @@ def update(request, participant_id):
                 messages.warning(request, 'Error sending e-mail. But a notification has been created on this website.')
     return redirect(request.POST.get('next', reverse('coursereg:index')))
 
+@require_POST
 @login_required
 def approve_all(request):
-    assert request.method == 'POST'
     if request.POST['origin'] == 'adviser':
         student = models.User.objects.get(id=request.POST['student_id'])
-        assert student.adviser == request.user
+        if not student.adviser == request.user: raise PermissionDenied
         if models.Participant.objects.filter(user=student, is_adviser_approved=False):
             student.is_dcc_review_pending = True
             student.save()
@@ -148,17 +151,18 @@ def approve_all(request):
         models.Participant.objects.filter(course__auto_instructor_approve=True, user=student).update(is_instructor_approved=True)
     elif request.POST['origin'] == 'instructor':
         course = models.Course.objects.get(id=request.POST['course_id'])
-        assert models.Participant.objects.filter(course=course,
-            user=request.user, participant_type=models.Participant.PARTICIPANT_INSTRUCTOR)
+        if not models.Participant.objects.filter(course=course,
+                user=request.user, participant_type=models.Participant.PARTICIPANT_INSTRUCTOR):
+            raise PermissionDenied
         models.Participant.objects.filter(course=course, is_adviser_approved=True).update(is_instructor_approved=True)
     return redirect(request.POST.get('next', reverse('coursereg:index')))
 
+@require_POST
 @login_required
 def delete(request, participant_id):
-    assert request.method == 'POST'
     participant = models.Participant.objects.get(id=participant_id)
-    assert participant.user == request.user
-    assert not participant.is_adviser_approved
-    assert not participant.course.is_last_reg_date_passed()
+    if not participant.user == request.user: raise PermissionDenied
+    if participant.is_adviser_approved: raise PermissionDenied
+    if participant.course.is_last_reg_date_passed(): raise PermissionDenied
     participant.delete()
     return redirect(request.GET.get('next', reverse('coursereg:index')))
